Refac/product list

+++ src/main/java/com/example/jejuairbnb/adminController/AdminProductDto/UpdateProductDto/UpdateProductRequestDto.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class UpdateProductRequestDto {
     private Long id;
     private String name;
-    private Long price;
+    private int price;
     private String img;
 }
